+++ src/controls/filePicker/IFilePickerProps.ts
@@ -28,12 +28,12 @@ export interface IFilePickerProps {
   /**
    * Handler when the file has been selected
    */
-  onSave:(filePickerResult: IFilePickerResult)=>void;
+  onSave:(filePickerResult: IFilePickerResult[])=>void;
 
   /**
    * Handler when file has been changed.
    */
-  onChange?: (filePickerResult: IFilePickerResult) => void;
+  onChange?: (filePickerResult: IFilePickerResult[]) => void;
 
   /**
    * Current context.
@@ -69,6 +69,11 @@ export interface IFilePickerProps {
    * Specifies if the picker button is disabled
    */
   disabled?: boolean;
+
+  /**
+   * Specifies if the picker button is hidden (if hidden, panel visibility can still be controlled with isPanelOpen)
+   */
+  hidden?: boolean;
 
   /**
    * Number of itmes to obtain when executing REST queries. Default 100.
@@ -110,6 +115,11 @@ export interface IFilePickerProps {
    */
   hideLocalUploadTab?: boolean;
 
+ /**
+   * Specifies if LocalMultipleUploadTab should be hidden.
+   */
+  hideLocalMultipleUploadTab?: boolean;
+
   /**
    * Specifies if LinkUploadTab should be hidden.
    */
@@ -137,6 +147,11 @@ export interface IFilePickerProps {
    * Optional additional renderer for Upload tab
    */
   renderCustomUploadTabContent?: (filePickerResult: IFilePickerResult) => JSX.Element | null;
+
+  /**
+   * Optional additional renderer for Multiple Upload tab
+   */
+  renderCustomMultipleUploadTabContent?: (filePickerResult: IFilePickerResult[]) => JSX.Element | null;
 
   /**
    * Specifies if Site Pages library to be visible on Sites tab