+++ src/controls/filePicker/MultipleUploadFilePickerTab/IMultipleUploadFilePickerTabState.ts
@@ -1,0 +1,7 @@
+import { IFilePickerResult } from "../FilePicker.types";
+
+export interface IMultipleUploadFilePickerTabState {
+  filePickerResult: IFilePickerResult[];
+  filePreview?: string;
+  // filesResult?: any;
+}